oprava toho co jsem podělal
